library_mgmt_swagger_full/requirements.txt
Django>=5.0,<6.0
djangorestframework>=3.15
djangorestframework-simplejwt>=5.3
drf-yasg>=1.21
django-filter>=24.2
python-dotenv>=1.0
